Code/requirements.txt
sqlalchemy
pandas
fbprophet
plotly
pymysql
pytz
twisted
scrapy
beautifulsoup4
requests
APScheduler
pyyaml
python-dateutil